requirements.txt
asgiref==3.8.1
Django==5.1.5
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
load-dotenv==0.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
